Add rust-toolchain file for sgx-utils repository
The new version of cargo that we have in the toolchain repository
supports overriding the toolchain version on a per-directory-tree
basis with a rust-toolchain file. Our toolchain has 1.18 as the default
Rust compiler version, so we need overrides for components that need
to build with the nightly compiler. This adds it for the sgx-utils
repository.

I tested that sgxs-tools and libenclave-tools build with the override.
Libenclave-tools won't build without the override, since it's using
features unsupported by 1.18.