Fix exit bug
`fn exit()` calling `abort_all_threads()` created a race condition:
return future in `fn syscall_loop()` would detect the exit and break
the syscall loop by receiving either the EnclaveAbort from the thread
that caused the exit or an EnclaveAbort::Secondary from other threads
that observed the exit in `fn dispatch()`.

To fix the issue:
- Avoid calling `abort_all_threads()` in `fn exit`. Do that only
  after the syscall loop is broken.
- Ignore EnclaveAbort::Secondary from non-main threads in return
  future which is not strictly needed after the previous change,
  but would make the code easier to read.

Also, when a non-main thread calls `std::process::exit()` with any
exit status, it should terminate the enclave execution. This was
not handled correctly for exit status 0.