Add an OID for attestation transparency. HRST-32
Attestation transparency information will be added to the
certificate used for joining the cluster, so we need an
OID for that purpose.

This is being added on the 1.x branch because roche
currently must be built with a 1.x version of sgx_pkix
due to dependencies on dcap-ql version 3.3. dcap-ql
can't be updated without updating mbedtls.