X509 OID for attestation transparency for sgx_pkix. HRST-33.
This adds an X509 OID for attestation transparency. This
change was previously made ont he sgx_pkix 0.1.x branch
(for sgx_pkix version 0.1.4). This change adds the new OID
on the master branch for sgx_pkix version 0.2.2, for when
we migrate roche to the latest releases of the crates
that belong to rust-sgx.