ci: add workflow for publish crate
Add a workflow to automatically publish crate.
This workflow is used for publish crate to crates.io.
The workflow can be manually triggered or by creating new release, when
the workflow is triggered by creating new release, please ensure the tag
for  the release follow this format: '${crate name}_v{version number}' .